Add bring.ts for content script and refactor vault timeout
