+++ src/background/redux/last-activity-time/selectors.ts
@@ -1,0 +1,4 @@
+import { RootState } from 'typesafe-actions';
+
+export const selectVaultLastActivityTime = (state: RootState): number | null =>
+  state.lastActivityTime;
