+++ src/background/redux/root-selector.ts
@@ -1,5 +1,6 @@
 export * from './active-origin/selectors';
 export * from './keys/selectors';
+export * from './last-activity-time/selectors';
 export * from './login-retry-count/selectors';
 export * from './login-retry-lockout-time/selectors';
 export * from './session/selectors';
