+++ src/background/redux/sagas/vault-sagas.ts
@@ -1,6 +1,5 @@
 import { put, select, takeLatest } from 'redux-saga/effects';
 import { getType } from 'typesafe-actions';
-import { alarms } from 'webextension-polyfill';
 
 import { getUrlOrigin } from '@src/utils';
 
@@ -25,6 +24,7 @@ import { Account } from '@libs/types/account';
 import { accountInfoReset } from '../account-info/actions';
 import { keysUpdated } from '../keys/actions';
 import { lastActivityTimeRefreshed } from '../last-activity-time/actions';
+import { selectVaultLastActivityTime } from '../last-activity-time/selectors';
 import { loginRetryCountReseted } from '../login-retry-count/actions';
 import {
   encryptionKeyHashCreated,
@@ -234,35 +234,47 @@ function* unlockVaultSaga(action: ReturnType<typeof unlockVault>) {
 }
 
 /**
- * This saga function is responsible for managing the vault timeout and locking mechanism.
- * It checks if the vault exists and is not locked, retrieves the vault timeout duration setting,
- * calculates the timeout duration value based on the setting, and creates an alarm to lock the vault.
- * If an error occurs during the execution, it logs the error.
+ * Saga to handle the timeout and locking mechanism of a vault based on its last activity time and a specified timeout duration setting.
+ *
+ * The generator function calculates the time elapsed since the last activity of the vault.
+ * If the vault exists, it is not locked, and the last
+ * activity time is available, it checks if the elapsed time surpasses the timeout duration.
+ * If true, it triggers a vault lock action
+ * immediately.
+ * If false, it sets up a delay for the remaining time until the timeout duration is met before locking the vault.
+ *
  */
 function* timeoutCounterSaga() {
+  const delay = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
+
   try {
-    // Check if the vault exists and is not locked
     const vaultDoesExist = yield* sagaSelect(selectVaultCipherDoesExist);
     const vaultIsLocked = yield* sagaSelect(selectVaultIsLocked);
-
-    // Get the vault timeout duration setting
+    const vaultLastActivityTime = yield* sagaSelect(
+      selectVaultLastActivityTime
+    );
     const vaultTimeoutDurationSetting = yield* sagaSelect(
       selectTimeoutDurationSetting
     );
-
-    // Calculate the timeout duration value based on the setting
     const timeoutDurationValue =
       MapTimeoutDurationSettingToValue[vaultTimeoutDurationSetting];
 
-    // If the vault exists and is not locked, create an alarm to lock the vault
-    if (vaultDoesExist && !vaultIsLocked) {
-      alarms.create('vaultLock', {
-        delayInMinutes: timeoutDurationValue
-      });
+    if (vaultDoesExist && !vaultIsLocked && vaultLastActivityTime) {
+      const currentTime = Date.now();
+      const timeoutExpired =
+        currentTime - vaultLastActivityTime >= timeoutDurationValue;
+
+      if (timeoutExpired) {
+        yield put(lockVault());
+      } else {
+        yield* sagaCall(delay, timeoutDurationValue);
+        yield put(lockVault());
+      }
     }
   } catch (err) {
-    // Log any errors that occur during the execution of the saga
-    console.error(err, 'err');
+    console.error(err);
+  } finally {
+    //
   }
 }
