+++ src/manifest.v2.json
@@ -11,7 +11,6 @@
     "storage",
     "tabs",
     "declarativeNetRequest",
-    "alarms",
     "https://image-proxy-cdn.make.services/*",
     "https://casper-assets.s3.amazonaws.com/*",
     "https://node.cspr.cloud/*",
